+++ crates/editor_ui/src/lib.rs
@@ -42,6 +42,15 @@ pub mod ui_registration;
 /// This module contains UI logic for view game camera image
 pub mod camera_view;
 
+/// UI plugin and common systems
+pub mod ui_plugin;
+
+/// Camera plugin and logic
+pub mod camera_plugin;
+
+///Selection logic
+pub mod selection;
+
 use bevy_debug_grid::{Grid, GridAxis, SubGrid, TrackedGrid, DEFAULT_GRID_ALPHA};
 use bevy_mod_picking::{
     backends::raycast::RaycastPickable,
@@ -83,6 +92,9 @@ use space_shared::{
 use space_undo::{SyncUndoMarkersPlugin, UndoPlugin, UndoSet};
 use ui_registration::BundleReg;
 
+use camera_plugin::*;
+use ui_plugin::*;
+
 use self::{mouse_check::MouseCheck, tools::gizmo::GizmoToolPlugin};
 
 pub const LAST_RENDER_LAYER: u8 = RenderLayers::TOTAL_LAYERS as u8 - 1;
@@ -99,10 +111,11 @@ pub mod prelude {
     pub use space_prefab::prelude::*;
     pub use space_shared::prelude::*;
 
+    pub use crate::camera_plugin::*;
+    pub use crate::selection::*;
     pub use crate::simple_editor_setup;
+    pub use crate::ui_plugin::*;
     pub use crate::EditorPlugin;
-    pub use crate::EditorUiAppExt;
-    pub use crate::EditorUiRef;
 }
 
 /// External dependencies for editor crate
@@ -138,11 +151,11 @@ impl PluginGroup for EditorPluginGroup {
             .add(EditorDefaultBundlesPlugin)
             .add(EditorDefaultCameraPlugin)
             .add(bevy_egui::EguiPlugin)
-            .add(EventListenerPlugin::<SelectEvent>::default())
+            .add(EventListenerPlugin::<selection::SelectEvent>::default())
             .add(DefaultInspectorConfigPlugin);
         res = EditorUiPlugin::default().add_plugins_to_group(res);
         res.add(PanOrbitCameraPlugin)
-            .add(EditorPickingPlugin)
+            .add(selection::EditorPickingPlugin)
             .add(bevy_debug_grid::DebugGridPlugin::without_floor_grid())
             .add(
                 WorldInspectorPlugin::default()
@@ -159,52 +172,6 @@ impl Plugin for EditorDefaultBundlesPlugin {
     fn build(&self, app: &mut App) {
         ui_registration::register_mesh_editor_bundles(app);
         ui_registration::register_light_editor_bundles(app);
-    }
-}
-
-pub struct EditorDefaultCameraPlugin;
-
-impl Plugin for EditorDefaultCameraPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_systems(
-            Update,
-            reset_editor_camera_state
-                .in_set(EditorSet::Editor)
-                .before(UiSystemSet),
-        );
-        app.add_systems(
-            Update,
-            update_pan_orbit
-                .after(reset_editor_camera_state)
-                .before(PanOrbitCameraSystemSet)
-                .in_set(EditorSet::Editor),
-        );
-        app.add_systems(
-            Update,
-            ui_camera_block
-                .after(reset_editor_camera_state)
-                .before(update_pan_orbit)
-                .in_set(EditorSet::Editor),
-        );
-    }
-}
-
-pub struct EditorPickingPlugin;
-
-impl Plugin for EditorPickingPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_plugins(bevy_mod_picking::DefaultPickingPlugins);
-
-        app.world
-            .resource_mut::<bevy_mod_picking::backends::raycast::RaycastBackendSettings>()
-            .require_markers = true;
-
-        app.add_systems(
-            PostUpdate,
-            (auto_add_picking, select_listener.after(UiSystemSet))
-                .run_if(in_state(EditorState::Editor)),
-        );
-        app.add_systems(PostUpdate, auto_add_picking_dummy);
     }
 }
 
@@ -245,14 +212,6 @@ impl Plugin for EditorSetsPlugin {
     }
 }
 
-/// This event used for selecting entities
-#[derive(Event, Clone, EntityEvent)]
-struct SelectEvent {
-    #[target]
-    e: Entity,
-    event: ListenerInput<Pointer<Down>>,
-}
-
 /// Allow editor manipulate GizmoConfig
 pub struct EditorGizmoConfigPlugin;
 
@@ -271,97 +230,12 @@ fn game_gizmos(mut gizmos_config: ResMut<GizmoConfig>) {
     gizmos_config.render_layers = RenderLayers::layer(0)
 }
 
-fn auto_add_picking(
-    mut commands: Commands,
-    query: Query<Entity, (With<PrefabMarker>, Without<Pickable>)>,
-) {
-    for e in query.iter() {
-        commands
-            .entity(e)
-            .insert(PickableBundle::default())
-            .insert(On::<Pointer<Down>>::send_event::<SelectEvent>())
-            .insert(RaycastPickable);
-    }
-}
-
 type AutoAddQueryFilter = (
     Without<PrefabMarker>,
     Without<Pickable>,
     With<Parent>,
     Changed<Handle<Mesh>>,
 );
-
-//Auto add picking for each child to propagate picking event up to prefab entitiy
-fn auto_add_picking_dummy(
-    mut commands: Commands,
-    query: Query<(Entity, &Handle<Mesh>), AutoAddQueryFilter>,
-    meshs: Res<Assets<Mesh>>,
-) {
-    for (e, mesh) in query.iter() {
-        //Only meshed entity need to be pickable
-        if let Some(mesh) = meshs.get(mesh) {
-            if mesh.primitive_topology() == PrimitiveTopology::TriangleList {
-                commands
-                    .entity(e)
-                    .insert(PickableBundle::default())
-                    .insert(RaycastPickable);
-            }
-        }
-    }
-}
-
-fn select_listener(
-    mut commands: Commands,
-    query: Query<Entity, With<Selected>>,
-    mut events: EventReader<SelectEvent>,
-    pan_orbit_state: ResMut<EditorCameraEnabled>,
-    keyboard: Res<Input<KeyCode>>,
-) {
-    if !pan_orbit_state.0 {
-        return;
-    }
-    for event in events.read() {
-        info!("Select Event: {:?}", event.e);
-        match event.event.button {
-            PointerButton::Primary => {
-                commands.entity(event.e).insert(Selected);
-                if !keyboard.pressed(KeyCode::ShiftLeft) {
-                    for e in query.iter() {
-                        commands.entity(e).remove::<Selected>();
-                    }
-                }
-            }
-            PointerButton::Secondary => { /*Show context menu?*/ }
-            PointerButton::Middle => {}
-        }
-    }
-}
-
-fn delete_selected(
-    mut commands: Commands,
-    query: Query<Entity, With<Selected>>,
-    keyboard: Res<Input<KeyCode>>,
-) {
-    let shift = keyboard.any_pressed([KeyCode::ShiftLeft, KeyCode::ShiftRight]);
-    let ctrl = keyboard.any_pressed([KeyCode::ControlLeft, KeyCode::ControlRight]);
-    let delete = keyboard.just_pressed(KeyCode::Back) || keyboard.just_pressed(KeyCode::Delete);
-
-    if ctrl && shift && delete {
-        for entity in query.iter() {
-            info!("Delete Entity: {entity:?}");
-            commands.entity(entity).despawn_recursive();
-        }
-    }
-}
-
-impl From<ListenerInput<Pointer<Down>>> for SelectEvent {
-    fn from(value: ListenerInput<Pointer<Down>>) -> Self {
-        Self {
-            e: value.target(),
-            event: value,
-        }
-    }
-}
 
 fn save_prefab_before_play(mut editor_events: EventWriter<space_shared::EditorEvent>) {
     editor_events.send(space_shared::EditorEvent::Save(
@@ -400,523 +274,9 @@ fn clear_and_load_on_start(
     }
 }
 
-/// Resource, which contains state for editor camera (default or any)
-#[derive(Resource, Default)]
-pub struct EditorCameraEnabled(pub bool);
-
-/// This system executes before all UI systems and is used to enable pan orbit camera on frame start
-pub fn reset_editor_camera_state(mut state: ResMut<EditorCameraEnabled>) {
-    *state = EditorCameraEnabled(true);
-}
-
-/// This system executes after all UI systems and is used to set pan orbit camera state.
-/// For example, it will block pan orbit camera if pointer is used by egui
-pub fn update_pan_orbit(
-    mut pan_orbit_query: Query<&mut PanOrbitCamera>,
-    state: Res<EditorCameraEnabled>,
-) {
-    for mut pan_orbit in pan_orbit_query.iter_mut() {
-        pan_orbit.enabled = state.0;
-    }
-}
-
-/// Sytem to block camera control if egui is using mouse
-pub fn ui_camera_block(
-    mut ctxs: Query<&mut EguiContext, With<PrimaryWindow>>,
-    mut state: ResMut<EditorCameraEnabled>,
-    game_view: Res<GameViewTab>,
-) {
-    let Ok(mut ctx_ref) = ctxs.get_single_mut() else {
-        return;
-    };
-    let ctx = ctx_ref.get_mut();
-    if ctx.is_using_pointer() || ctx.is_pointer_over_area() {
-        let Some(pos) = ctx.pointer_latest_pos() else {
-            return;
-        };
-        if let Some(area) = game_view.viewport_rect {
-            if area.contains(pos) {
-            } else {
-                *state = EditorCameraEnabled(false);
-            }
-        } else {
-            *state = EditorCameraEnabled(false);
-        }
-    }
-}
-
-type ChangeCameraQueryFilter = (Without<EditorCameraMarker>, With<CameraPlay>);
-
-/// System to change camera from editor camera to game camera (if exist)
-pub fn change_camera_in_play(
-    mut cameras: Query<&mut Camera, (With<EditorCameraMarker>, Without<CameraPlay>)>,
-    mut play_cameras: Query<(&mut Camera, &CameraPlay), ChangeCameraQueryFilter>,
-) {
-    if !play_cameras.is_empty() {
-        let (mut some_camera, _) = play_cameras.iter_mut().next().unwrap();
-        cameras.single_mut().is_active = false;
-        some_camera.is_active = true;
-    }
-}
-
-/// System to change camera from game camera to editor camera (if exist)
-pub fn change_camera_in_editor(
-    mut cameras: Query<&mut Camera, With<EditorCameraMarker>>,
-    mut play_cameras: Query<&mut Camera, Without<EditorCameraMarker>>,
-) {
-    for mut ecam in cameras.iter_mut() {
-        ecam.is_active = true;
-    }
-
-    for mut play_cam in play_cameras.iter_mut() {
-        play_cam.is_active = false;
-    }
-}
-
 pub trait FlatPluginList {
     fn add_plugins_to_group(&self, group: PluginGroupBuilder) -> PluginGroupBuilder;
 }
-
-///Camera with this component will not be disabled in Editor state
-#[derive(Component)]
-pub struct DisableCameraSkip;
-
-fn disable_no_editor_cams(
-    mut cameras: Query<&mut Camera, (Without<DisableCameraSkip>, Without<EditorCameraMarker>)>,
-) {
-    for mut cam in cameras.iter_mut() {
-        cam.is_active = false;
-    }
-}
-
-#[derive(Component)]
-pub struct NotShowCamera;
-
-fn draw_camera_gizmo(
-    mut gizmos: Gizmos,
-    cameras: Query<
-        (&GlobalTransform, &Projection),
-        (
-            With<Camera>,
-            Without<EditorCameraMarker>,
-            Without<DisableCameraSkip>,
-            Without<NotShowCamera>,
-        ),
-    >,
-) {
-    for (transform, _projection) in cameras.iter() {
-        let transform = transform.compute_transform();
-        let cuboid_transform = transform.with_scale(Vec3::new(1.0, 1.0, 2.0));
-        gizmos.cuboid(cuboid_transform, Color::PINK);
-
-        let scale = 1.5;
-
-        gizmos.line(
-            transform.translation,
-            transform.translation
-                + transform.forward() * scale
-                + transform.up() * scale
-                + transform.right() * scale,
-            Color::PINK,
-        );
-        gizmos.line(
-            transform.translation,
-            transform.translation + transform.forward() * scale - transform.up() * scale
-                + transform.right() * scale,
-            Color::PINK,
-        );
-        gizmos.line(
-            transform.translation,
-            transform.translation + transform.forward() * scale + transform.up() * scale
-                - transform.right() * scale,
-            Color::PINK,
-        );
-        gizmos.line(
-            transform.translation,
-            transform.translation + transform.forward() * scale
-                - transform.up() * scale
-                - transform.right() * scale,
-            Color::PINK,
-        );
-
-        let rect_transform = Transform::from_xyz(0.0, 0.0, -scale);
-        let rect_transform = transform.mul_transform(rect_transform);
-
-        gizmos.rect(
-            rect_transform.translation,
-            rect_transform.rotation,
-            Vec2::splat(scale * 2.0),
-            Color::PINK,
-        );
-    }
-}
-
-/// All systems for editor ui wil be placed in UiSystemSet
-#[derive(SystemSet, Hash, PartialEq, Eq, Debug, Clone, Copy)]
-pub struct UiSystemSet;
-
-/// Plugin for editor ui
-pub struct EditorUiPlugin {
-    pub use_standard_layout: bool,
-}
-
-impl Default for EditorUiPlugin {
-    fn default() -> Self {
-        Self {
-            use_standard_layout: true,
-        }
-    }
-}
-
-/// State to determine if editor ui should be shown (ot hidden for any reason)
-#[derive(Hash, PartialEq, Eq, Debug, Clone, States, Default)]
-pub enum ShowEditorUi {
-    #[default]
-    Show,
-    Hide,
-}
-
-impl FlatPluginList for EditorUiPlugin {
-    fn add_plugins_to_group(&self, group: PluginGroupBuilder) -> PluginGroupBuilder {
-        let mut res = group
-            .add(SelectedPlugin)
-            .add(EditorUiCore::default())
-            .add(GameViewPlugin)
-            .add(bottom_menu::BottomMenuPlugin)
-            .add(MouseCheck)
-            .add(CameraViewTabPlugin)
-            .add(SpaceHierarchyPlugin::default())
-            .add(SpaceInspectorPlugin)
-            .add(GizmoToolPlugin)
-            .add(ChangeChainViewPlugin)
-            .add(settings::SettingsWindowPlugin);
-
-        if self.use_standard_layout {
-            res = res.add(DefaultEditorLayoutPlugin);
-        }
-
-        res
-    }
-}
-
-impl PluginGroup for EditorUiPlugin {
-    fn build(self) -> PluginGroupBuilder {
-        let mut group = PluginGroupBuilder::start::<Self>();
-        group = self.add_plugins_to_group(group);
-        group
-    }
-}
-
-pub struct DefaultEditorLayoutPlugin;
-
-impl Plugin for DefaultEditorLayoutPlugin {
-    fn build(&self, app: &mut App) {
-        let mut editor = app.world.resource_mut::<EditorUi>();
-        editor.tree = egui_dock::DockState::new(vec![EditorTabName::GameView]);
-
-        let [_game, _inspector] = editor.tree.main_surface_mut().split_right(
-            egui_dock::NodeIndex::root(),
-            0.8,
-            vec![EditorTabName::Inspector],
-        );
-        let [_hierarchy, _game] =
-            editor
-                .tree
-                .main_surface_mut()
-                .split_left(_game, 0.2, vec![EditorTabName::Hierarchy]);
-    }
-}
-
-pub struct EditorUiCore {
-    pub disable_no_editor_cams: bool,
-}
-
-impl Default for EditorUiCore {
-    fn default() -> Self {
-        Self {
-            disable_no_editor_cams: true,
-        }
-    }
-}
-
-impl Plugin for EditorUiCore {
-    fn build(&self, app: &mut App) {
-        app.add_state::<ShowEditorUi>();
-
-        app.configure_sets(
-            Update,
-            UiSystemSet
-                .in_set(EditorSet::Editor)
-                .run_if(in_state(EditorState::Editor).and_then(in_state(ShowEditorUi::Show))),
-        );
-        app.init_resource::<EditorUi>();
-        app.init_resource::<ScheduleEditorTabStorage>();
-        app.add_systems(
-            Update,
-            (
-                show_editor_ui
-                    .before(update_pan_orbit)
-                    .before(ui_camera_block)
-                    .after(bottom_menu::menu),
-                set_camera_viewport,
-            )
-                .in_set(UiSystemSet),
-        );
-
-        app.add_systems(
-            PostUpdate,
-            set_camera_viewport
-                .run_if(has_window_changed)
-                .in_set(UiSystemSet),
-        );
-        app.add_systems(
-            Update,
-            reset_camera_viewport.run_if(in_state(EditorState::Game)),
-        );
-        app.add_systems(OnEnter(ShowEditorUi::Hide), reset_camera_viewport);
-        app.editor_tab_by_trait(EditorTabName::GameView, GameViewTab::default());
-
-        app.editor_tab_by_trait(
-            EditorTabName::Other("Debug World Inspector".to_string()),
-            self::debug_panels::DebugWorldInspector {},
-        );
-
-        app.init_resource::<EditorLoader>();
-
-        app.insert_resource(EditorCameraEnabled(true));
-
-        app.add_systems(
-            Startup,
-            (set_start_state, apply_state_transition::<EditorState>).chain(),
-        );
-
-        //play systems
-        app.add_systems(OnEnter(EditorState::GamePrepare), save_prefab_before_play);
-        app.add_systems(
-            OnEnter(SaveState::Idle),
-            to_game_after_save.run_if(in_state(EditorState::GamePrepare)),
-        );
-
-        app.add_systems(OnEnter(EditorState::Game), change_camera_in_play);
-
-        app.add_systems(
-            OnEnter(EditorState::Editor),
-            (clear_and_load_on_start, set_camera_viewport),
-        );
-
-        app.add_systems(
-            Update,
-            (draw_camera_gizmo, delete_selected)
-                .run_if(in_state(EditorState::Editor).and_then(in_state(ShowEditorUi::Show))),
-        );
-
-        if self.disable_no_editor_cams {
-            app.add_systems(
-                Update,
-                disable_no_editor_cams.run_if(in_state(EditorState::Editor)),
-            );
-
-            app.add_systems(OnEnter(EditorState::Editor), change_camera_in_editor);
-        }
-
-        app.add_event::<SelectEvent>();
-
-        app.init_resource::<BundleReg>();
-    }
-}
-
-/// This system use to show all egui editor ui on primary window
-/// Will be usefull in some specific cases to ad new system before/after this system
-pub fn show_editor_ui(world: &mut World) {
-    let Ok(egui_context) = world
-        .query_filtered::<&mut EguiContext, With<PrimaryWindow>>()
-        .get_single(world)
-    else {
-        return;
-    };
-    let mut egui_context = egui_context.clone();
-
-    world.resource_scope::<EditorUi, _>(|world, mut editor_ui| {
-        editor_ui.ui(world, egui_context.get_mut());
-    });
-}
-
-/// This resource contains registered editor tabs and current dock tree state
-#[derive(Resource)]
-pub struct EditorUi {
-    pub registry: HashMap<EditorTabName, EditorUiReg>,
-    pub tree: egui_dock::DockState<EditorTabName>,
-}
-
-impl Default for EditorUi {
-    fn default() -> Self {
-        Self {
-            registry: HashMap::default(),
-            tree: egui_dock::DockState::new(vec![]),
-        }
-    }
-}
-
-/// This enum determine how tab was registered.
-/// ResourceBased - tab will be registered as resource
-/// Schedule - tab will be registered as system
-pub enum EditorUiReg {
-    ResourceBased {
-        show_command: EditorTabShowFn,
-        title_command: EditorTabGetTitleFn,
-    },
-    Schedule,
-}
-
-impl EditorUi {
-    pub fn ui(&mut self, world: &mut World, ctx: &mut egui::Context) {
-        //collect tab names to vec to detect visible
-        let mut visible = vec![];
-        for (_surface_index, tab) in self.tree.iter_all_nodes() {
-            match tab {
-                egui_dock::Node::Empty => {}
-                egui_dock::Node::Leaf {
-                    rect: _,
-                    viewport: _,
-                    tabs,
-                    active: _,
-                    scroll: _,
-                } => visible.extend(tabs.clone()),
-                egui_dock::Node::Vertical {
-                    rect: _,
-                    fraction: _,
-                } => {}
-                egui_dock::Node::Horizontal {
-                    rect: _,
-                    fraction: _,
-                } => {}
-            }
-        }
-
-        let cell = world.as_unsafe_world_cell();
-
-        let mut command_queue = CommandQueue::default();
-        let mut commands = Commands::new(&mut command_queue, unsafe { cell.world() });
-
-        let mut tab_viewer = unsafe {
-            EditorTabViewer {
-                commands: &mut commands,
-                world: cell.world_mut(),
-                registry: &mut self.registry,
-                visible,
-                tab_commands: vec![],
-            }
-        };
-
-        DockArea::new(&mut self.tree)
-            .show_add_buttons(true)
-            .show_add_popup(true)
-            .show(ctx, &mut tab_viewer);
-
-        let windows_setting = unsafe { cell.world_mut().resource_mut::<NewWindowSettings>() };
-        for command in tab_viewer.tab_commands {
-            match command {
-                EditorTabCommand::Add {
-                    name,
-                    surface,
-                    node,
-                } => match windows_setting.new_tab {
-                    NewTabBehaviour::Pop => {
-                        self.tree.add_window(vec![name]);
-                    }
-                    NewTabBehaviour::SameNode => {
-                        if let Some(tree) = self
-                            .tree
-                            .get_surface_mut(surface)
-                            .and_then(|surface| surface.node_tree_mut())
-                        {
-                            tree.set_focused_node(node);
-                            tree.push_to_focused_leaf(name);
-                        }
-                    }
-                    NewTabBehaviour::SplitNode => {
-                        if let Some(surface) = self.tree.get_surface_mut(surface) {
-                            surface
-                                .node_tree_mut()
-                                .unwrap()
-                                .split_right(node, 0.5, vec![name]);
-                        }
-                    }
-                },
-            }
-        }
-
-        unsafe {
-            command_queue.apply(cell.world_mut());
-        }
-    }
-}
-
-/// Trait for registering editor tabs via app.**
-pub trait EditorUiAppExt {
-    fn editor_tab_by_trait<T>(&mut self, tab_id: EditorTabName, tab: T) -> &mut Self
-    where
-        T: EditorTab + Resource + Send + Sync + 'static;
-    fn editor_tab<T>(
-        &mut self,
-        tab_id: EditorTabName,
-        title: egui::WidgetText,
-        tab_systesm: impl IntoSystemConfigs<T>,
-    ) -> &mut Self;
-}
-
-impl EditorUiAppExt for App {
-    fn editor_tab_by_trait<T>(&mut self, tab_id: EditorTabName, tab: T) -> &mut Self
-    where
-        T: EditorTab + Resource + Send + Sync + 'static,
-    {
-        self.insert_resource(tab);
-        let show_fn = Box::new(
-            |ui: &mut egui::Ui, commands: &mut Commands, world: &mut World| {
-                world.resource_scope(|scoped_world, mut data: Mut<T>| {
-                    data.ui(ui, commands, scoped_world)
-                });
-            },
-        );
-        let reg = EditorUiReg::ResourceBased {
-            show_command: show_fn,
-            title_command: Box::new(|world| world.resource_mut::<T>().title()),
-        };
-
-        self.world
-            .resource_mut::<EditorUi>()
-            .registry
-            .insert(tab_id, reg);
-        self
-    }
-
-    fn editor_tab<T>(
-        &mut self,
-        tab_id: EditorTabName,
-        title: egui::WidgetText,
-        tab_systesm: impl IntoSystemConfigs<T>,
-    ) -> &mut Self {
-        let mut tab = ScheduleEditorTab {
-            schedule: Schedule::default(),
-            title,
-        };
-
-        tab.schedule.add_systems(tab_systesm);
-
-        self.world
-            .resource_mut::<ScheduleEditorTabStorage>()
-            .0
-            .insert(tab_id.clone(), tab);
-        self.world
-            .resource_mut::<EditorUi>()
-            .registry
-            .insert(tab_id, EditorUiReg::Schedule);
-        self
-    }
-}
-
-/// Temporary resource for pretty system, based tab registration
-pub struct EditorUiRef(pub egui::Ui);
 
 /// This method prepare default lights and camera for editor UI. You can create own conditions for your editor and use this method how example
 pub fn simple_editor_setup(mut commands: Commands) {
